v4l::device

Struct Device

pub struct Device { /* private fields */ }

Linux capture device abstraction

Implementations

impl Device

pub fn new(index: usize) -> Result<Self>

Returns a capture device by index

Devices are usually enumerated by the system. An index of zero thus represents the first device the system got to know about.

Arguments

index - Index (0: first, 1: second, ..)

Example

use v4l::device::Device;
let dev = Device::new(0);

pub fn with_path<P: AsRef<Path>>(path: P) -> Result<Self>

Returns a capture device by path

Linux device nodes are usually found in /dev/videoX or /sys/class/video4linux/videoX.

Arguments

path - Path (e.g. “/dev/video0”)

Example

use v4l::device::Device;
let dev = Device::with_path("/dev/video0");

pub fn handle(&self) -> Arc<Handle>

Returns the raw device handle

pub fn query_caps(&self) -> Result<Capabilities>

Returns video4linux framework defined information such as card, driver, etc.

pub fn query_controls(&self) -> Result<Vec<Description>>

Returns the supported controls for a device such as gain, focus, white balance, etc.

pub fn control(&self, id: u32) -> Result<Control>

Returns the control value for an ID

Arguments

id - Control identifier

pub fn set_control(&self, ctrl: Control) -> Result<()>

Modifies the control value

Arguments

ctrl - Control to be set

pub fn set_controls(&self, ctrls: Vec<Control>) -> Result<()>

Modifies the control values atomically

Arguments

ctrls - Vec of the controls to be set